EventSource: Add version ID setter/getters
The version ID will be used as part of the synchronization process of
event-notifications which SubscriptionManager will implement:
- At the beginning of a Client connection, the client database might
contain old copies of a huge amount of EventSource objects, which were
received in a previous connection.
- To synchronize the client with the WOT database, WOT will send all
current version of all EventSource objects. The client then will have to
delete the ones which were not included in the most recent dataset. This
can be done efficiently by deleting all objects whose versionID is not
the one of the current dataset.