SubscriptionManager: Replace SynchronizationContainer with BeginSynchronizationNotification
This is part of replacing the whole synchronization mechanism with
something different:

The previous concept of SynchronzitationContainer was to store
Subscription synchronizations as huge, bulk data Bucket containing ALL
EventSource objects part of the synchronization. This would be difficult
to implement as fred Buckets are not easily made persistent across
restarts. Also, it would have been difficult to garbage collect them.

Instead, we will store each EventSource object part of the
synchronization in a single ObjectChangedNotification object in the
database, just as we do with regular changes of bjects. So instead of 1
huge object, we will send N small notifications to the client
But to make the client able to delete all old EventSource objects from
its database which existed before the synchronization, it will have to
know which they are: If the synchronization is sent one-by-one-object
isntead of a bulk dataset, the client will never have the bulk dataset
at hand at one time to compare it to the preexisting data.
So to allow the client to garbage-collect the pre-existing data, we give
the client the opportunity to mark-and-sweep: We give each
synchronization a UUID versionID. All objects part of the
synchronization will have the same UUID. Thus, if the client stores the
UUID for each received EventSource, after the synchronization has ended,
it only has to delete all objects with a different UUID to get rid of
any pre-existing data.

Notice:
- This only adds class BeginSynchronizationNotification, the
symmetrical EndSynchronizationNotification will be added by the
following commit.
- Also, the JavaDoc of the whole of SubscriptionManager
is not adapted. It is due a full JavaDoc  review anyway, so I  will deal
with this once the branch use-plugin-fcp-rewrite is finished (the
current branch split-up-subscription-synchronization is a sub-branch of
that and will be merged into it before finishing
use-plugin-fcp-rewrite).
- The actual implementation of the mechanism using the new class will be
done in following commits.