SubscriptionManager: Fix ambiguous JavaDoc
- With regards to SubscriptionManager, synchronization is a technical
term which has a different meaning than thread synchronization so we
should use a different word when talking about threads there. Thereby,
this commit uses the noun "thread safety" instead of "thread
synchronization" and the verb "to lock" instead of "to synchronize".
- Also, while I was at it, I unified the HTML of the JavaDoc and added a
link to Persistent.transactionLock() instead of only talking about "the
database transaction lock"