SubscriptionManager.Subscription: Get rid of synchronizeSubscriberByFCP()
- The special case of sending the Identity/Trust/Score objects in bulk
there was  replaced by the mechanism of
storeSynchronizationWithoutCommit(): It stores the synchronization as a
series of regular ObjectChangedNotifcation, which can be deployed by the
regular notifySubscriberByFCP().
- Consequently, the parts of the JavaDoc which were better than at
notifySubscriberByFCP() are moved to that one.
- What remains is adapting notifySubscriberByFCP() to be able to also
deploy the special BeginSynchronizationNotification and
EndSynchronizationNotification. They are rather simple, so they didn't
justify keeping the synchronizeSubscriberByFCP() to contain them.