SubscriptionManager.storeNewSubscriptionAndCommit(): Fix doing database modifications outside of a transaction
- It previously was assumed that this function would not have to be
called while taking the database lock: All modifications to the database
it did were inside of Subscription.storeAndCommit(), so it was enough
that that function did take care of doing a transaction. But recently
storeNewSubscriptionAndCommit() was changed to call
Subscription.storeSynchronizationWithoutCommit(), which does modify the
database. So now the caller of the storeNewSubscriptionAndCommit() must
take sure to lock the database.
- The JavaDoc is fixed to mention the above.
- The function is renamed from "...AndCommit()" to "...WithoutCommit()"
to stress that outside transaction management is needed. The committing
is removed from the function of course.
- The Subscription.storeAndCommit() is removed since it is not used
anymore.
- The callers are fixed to do transaction management.