Identity/Trust/Score: Implement storage of version ID
- As required by the interface EventSource which they already implement.
- Notice: Even though this commit implements storage to db4o, it is not
actually used by WOT: WOT will only call the setVersionID() upon
temporary clones which are not stored in db4o. Nevertheless, I decided
that implementing storage is a good idea: The primary purpose of the
value is to allow event-notifications client apps to mark-and-sweep
garbage collect obsolete Identity/Trust/Score objects. For that, they
obviously need storage. And they are encouraged to copy paste the WOT
Identity/Trust/Score classes (see FCPClientReferenceImplementation),
so its a good idea to provide an implementation which will work at the
client side as well. So  we make the  storage working already in case
client authors decide to just use the existing WOT db4o code for
storage.