SubscriptionManagerFCPTest: Adapt to new way of subscription synchronization
The code which this tests is:
- SubscriptionManager.storeNewSubscriptionWithoutCommit()
- the following callee of the above:
SubscriptionManager.Subscription.storeSynchronizationWithoutCommit()
- FCPInterface.sendBeginOrEndSynchronizationNotification()

Notice: This test will currently fail. This is because it actually
detects a bug in
FCPInterface.sendBeginOrEndSynchronizationNotification():
The type-casting to check the runtime type of generics won't work as
intended as the type is not known at runtime.
The following commits shall fix this.