SubscriptionManagerFCPTest: Fix to use new way of Subscription synchronization
Fixes testAllRandomized():
- The underlying subscribeAndSynchronize() is adapted to understand the
new FCP message layout of Subscription synchronization. The code is a
copy of parts of testSubscribeTo().
- Notice that I did amend the function to do all tests which
testSubscribeTo() does. For the old revision, I had decided to only test
whats relevant to the testAllRandomized(). But why not have the function
test as much as possible?
- All unit tests should succeed by now. Next step will be to adapt
FCPClientReferenceImplementation to work with the new Subscription
synchronization. Once that is done, the FIXME which is added by this
commit can be resolved: The importSynchronization() uses
FCPClientReferenceImplementation currently, so thats why this commit
does not use it yet and marks usage of it as a FIXME.