DebugFCPClient: Add getter for filtering the dataset by versionID
- These will be needed for
implementing BeginSubscriptionSynchronizationHandlerImpl and
EndSubscriptionSynchronizationHandlerImpl to check the dataset for
correctness.
- As this class is a debugging class, and subscription synchronization
typically only happens once per debug run (at the beginning of the
connection), it is OK to implement this in a slow way. The fast
implementation of having a separate datastructur for storing the
elements by their versionID would be more complex and thus more
difficult to understand for new developers. While there is no
non-debug version of this class, it still sort of aims to help new
developers understand how event-notifications is used and thus I want to
keep it simple.