SubscriptionManager: Don't send pending Notifications when subscribing
This is not necessary anymore: The class was recently changed to not
send out the synchronization immediately while subscribing anymore.
Instead, it stores it to the  Notification queue. And as the
Notification queue is deployed in order, everything will be fine: The
purpose of the removed code was to preserve the order.