FCPInterface: Unify *ChangedNotification into one message type
- This will allow simplification of the parser code in
FCPClientReferenceImplementation as described in the removed FIXME. That
code will appear in following commits.
- Use ObjectChangedEventNotification instead of
ObjectChangedNotification as we want to replace "Notification" with
"Event" as its more easy to understand. Its good to do this for the FCP
API now already as that will be the most difficult to change. See
https://bugs.freenetproject.org/view.php?id=5742
- Use the SubscriptionType enum of FCPClientReferenceImplementation as
that allows easy conversion of Java types to the Strings
"Identities" / "Trusts" / "Scores" used in FCP.