FCP: Rename "*SynchronizationNotification" to "*SynchronizationEvent"
The whole Notification stuff was advertised as "event-notifications"
during development a lot. To make it easy for developers to get used it,
"Notification" should be renamed to Event everywhere. Unfortunately,
this is a lot of work with regards to the class hierarchy. Still, we can
rename the stuff to "Event" in FCP message at least, as that is what the
client developers will get in contact with. Also, FCP messages would be
difficult to rename afterwards because it would break client parsers, so
its better to do this now before the code has been released.