FCPInterface.handleSubscribe(): Fix wrong JavaDoc with minimal effort
The backend of this function was recently changed a lot: The whole
mechanism of subscription synchronization was revamped as part of
split-up-subscription-synchronization.
This commit removes the incorrect obsolete JavaDoc and replaces it with
an updated version. The updated version is intentionally left short
though, because we want to get out a new release as soon as possible and
thus I don't want to spend too much time on documentation.
A TODO is added to improve the JavaDoc.

If you need to know how the whole mechanism of event Subscriptions
works, you can read FCPClientReferenceImplementation for a reference
user of it.