Add missing @Override annotations to almost everything
"Almost everything" = everything except
plugins.WebOfTrust.introduction.captcha.kaptcha.*
as that is a third party library.

Done using the Eclipse feature:
Right click on project / Source / Clean up ...

FYI: There is a bug in Eclipse which I had to work around:
If this feature modifies the whitespace of the following line after
@Override for no reason, that is because your whitespace settings in
Ecplise (whether to use tabs or spaces) do not match the whitespace
which was in the file.
For example if the function following the @Override is indented by tabs,
and Eclipse is configured to indent by spaces, then the function
indentation will be changed by the feature from tabs to spaces even
though it should only add the @Override.
To fix that behavior, reconfigure Eclipse to use the opposite of what
was configured. If spaces were configured, configure it to use tabs.
Notice that if you have a mixture of functions, some indented by tabs,
some by spaces, then this will make the opposite break then.
So to produce a commit which adds @Override for both, you would:
- First use the auto @Override adding feature with tabs as indentation
- Commit all changes which were correct (= those which only added the
@Override and did not change the following line) excluding the incorrect
ones, using "git commit --patch"
- Drop the remaining changes using "git reset --hard"
- Then reconfigure Eclipse to use spaces as indentation and run the
@Override adding again.
- Amend the commit to include the remaining changes using "git commit
--patch --amend".