FCPClientReferenceImplementation: Remove resolved FIXME
I've done a full review of the class to ensure basic correctness of the
code and existing JavaDoc. It could need some more new JavaDoc but we
want to get a new build out ASAP and decided to postget a new build out
ASAP and decided to postpone writing of new JavaDoc. So this commit adds
a TODO about adding more JavaDoc.
With regards to the test run which has been requested by the FIXME: I've
already done several small test runs and the code seems to work. I'll do
a larger full test run when the whole of the new build is complete. No
FIXME is needed for that IMHO as doing a long test run is a regular part
of deployment anyway.