SubscriptionManager: Speed up shutdown a lot
It was slow because  stop() waited for run() to finish (implicitely by
mTicker.shutdown(), and run() can take a very long time because it
deploys all existing event-Notifications:
- Event-Notifications are queued up to 60 seconds before they are send
out, so the queue can become large.
- They are being deployed by network to the client, which can take
a long time due to slow connections.
- They are deployed synchronously as they need to arrive in proper order
at the client. So it must wait for the client to process them.

Thus, this commit speeds shutdown up by making stop() do
Thread.interrupt() upon run(). This will cause the underlying
sendSynchronous() to throw InterruptedException, which will gracefully
abort sending of the Notification queue. It has a granularity of
checking the interrupted state for every single Notification, as we use
one call to sendSynchronous() per Notification.

Also, a TODO is added to improve fred to make doing this easier: We
could amend TrivialTicker to automatically interrupt() the currently
running job during shutdown.
As the TODO is not within fred itself, I will file a bug at fred to
make fred developers notice it.