FCPInterface: Resolve FIXME of proper shutdown of sendSynchronous()
The previous commit fb29a06086b01268753f7396723e43d12e41c84a changed
SubscriptionManager to do Thread.interrupt() upon all threads which call
the stuff in FCPInterface which would be candidates for interruption (=
callers of FCPPluginClient.sendSynchronous()). So FCPInterface doesn't
need to deal with this, and we can just remove the FIXME.
Still, this commit adds JavaDoc to the callers of sendSynchronous()
which states that interrupt() must be used to ensure that future changes
to stuff outside of FCPInterface respect the need for interrupt() during
shutdown.