SubscriptionManager.subscribeTo*(): Fix obsolete JavaDoc
The recent branch split-up-subscription-synchronization changed the way
subscription synchronization works to *not* transfer the synchronization
to the client while subscribing. Instead, during subscribe*(), the data
is merely copied from the main WOT database to the event Notification
queue. The Notifications will be deployed to the client in a different
thread in the future. See
Subscription.storeSynchronizationWithoutCommit().