DeleteOwnIdentityWizardPage: Synchronize
Synchronize around usage of the OwnIdentity object because it is stored
in the database. Clone it so we can keep using it across multiple
functions - synchronizing across all of them wouldn't be possible as
they are being called from the unrelated outside web interface code.

This is a workaround for the following issue:
https://bugs.freenetproject.org/view.php?id=6247