IdentityPage: Synchronize
This is a workaround for the following issue:
https://bugs.freenetproject.org/view.php?id=6247

Synchronize around usage of the (Own)Identity objects because they are
stored in the database and thus subject to 6247.

Notice: The approach of clone()ing the "Identity identity" in the
constructor which we do for the OwnIdentity in the constructor
won't work for the Identity:
The "Identity identity" is passed to getGivenTrusts() /
getReceivedTrusts() in make(), and those functions use
it in database queries. For the database queries to work, the very
same Identity object needs to be used as was obtained from a previous
database query. Thus, we cannot clone.
So instead, we query the Identity in make(). We use an ErrorPage there
for dealing with UnknownIdentityException, since make() may not throw -
as opposed to the constructor. Its a better place to handle this here
instead of throwing it in the constructor anyway: Throwing it in the
constructor would signal the caller that the logged in OwnIdentity did
not exist anymore - which is wrong in case of the non-own "Identity
identity".
The OwnIdentity is not queried in make() but in the constructor as it
was already - for the same reason: We better query it in the constructor
so we can throw there to signal the caller to log out the user since the
backing OwnIdentity does not exist anymore.