KnownIdentitiesPage: Synchronize
This is a workaround for the following issue:
https://bugs.freenetproject.org/view.php?id=6247

Synchronize around usage of the OwnIdentity object because it is stored
in the database and thus subject to 6247. Clone it so we can keep using
it across multiple functions - synchronizing across all of them wouldn't
be possible as they are being called from the unrelated outside web
interface code.
Also synchronize around all the other database quering code as well.

Notice: The OwnIdentity is not queried in make...() but in the
constructor as it was already - for the same reason: We better query it
in the constructor so we can throw there to signal the caller to log out
the user since the backing OwnIdentity does not exist anymore.