MyIdentityPagePage: Synchronize
This is a workaround for the following issue:
https://bugs.freenetproject.org/view.php?id=6247

Synchronize around usage of the OwnIdentity object because it is stored
in the database and thus subject to 6247. Clone it so we can keep using
it across multiple functions - synchronizing across all of them wouldn't
be possible as they are being called from the unrelated outside web
interface code.
Also synchronize around all the other database quering code as well:
Move it to the constructor so we only need to take the lock once.

Notice: The size() actually returns integers, not longs. The previous
code which used longs was wrong. Thus, use integers here.