FCPInterface.handleCreateIdentity(): Synchronize
This is a workaround for the following issue:
https://bugs.freenetproject.org/view.php?id=6247

Synchronize around usage of the OwnIdentity object because it is stored
in the database and thus subject to 6247.
Notice: The removed comment about preserving the locking order didn't
make any sense, it was probably obsolete - we only take one lock here,
so there is no order.