FCPInterface.handleGetTrusts(): Document why we synchronize
The TODO of removing the lock must IMHO be eliminated: We process a
list, and thus a database which is implemented optimally will read the
results one by one at the time of quering each result from the
iterator.
Thus, the returned data would not be coherent without synchronization.

If we ever move to a database which doesn't require this sync, the
getAllTrusts() will have to be rewritten probably as part of the
general database refactoring to contain database query code specific to
the new database. Then we will review its JavaDoc as part of the
rewrite anyway, notice that it says the caller has to synchronize, and
fix callers to not synchronize anymore.