DelayedBackgroundJob: Rephrase interface contract to match implementation
Changing the delay to be minimal instead of maximal sounds like a
disruptive change, but it is not:
The whole point behind developing BackgroundJobs was that WOT wants to
postpone certain computations with high base overhead (= processing time
required independent of size of input) which are triggered by only 1
input element having changed to wait for N elements to aggregate. This
causes the base overhead to be low compared to computation time for
processing the N elements.
Thus, the delay is a MINIMUM, not a maximum. The goal is NOT to execute
something as fast as possible, but to *wait* for a minimal amount of
time so elements can aggregate.

All current implementations of DelayedBackgroundJob already meet the
changed contract of being a minimal delay. The JavaDoc which is changed
by this commit was wrong basically.
Thus, their JavaDoc is also changed (and deduplicated).