SynchronousDelayedBackgroundJob: Fix wrong claim about TickerDelayedBackgroundJob
It also doesn't wait for the thread to exit, but merely for no more code
to be runable on it.
As a consequence, add the comment to its waitForTermination() as well.

(And move it to waitForTermination() for SynchronousDelayedBackgroundJob
since thats a better place for it because we're talking about
termination issues)