SynchronousDelayed..Test: Fix assertion failure
We deliberately call triggerExecution() in parallel on many threads.
The contract of triggerExecution() specifies that multiple calls will
not cause multiple executions if the jobs is still waiting to execute.
Thus not every one of the parallel calls to the function will cause an
execution, and the limit which we asserted to be equals is actually an
upper boundary.