TickerDelayedBackgroundJobTest: Fix AssertionError
java.lang.AssertionError: expected:<1> but was:<2>
	at org.junit.Assert.fail(Assert.java:88)
	at org.junit.Assert.failNotEquals(Assert.java:743)
	at org.junit.Assert.assertEquals(Assert.java:118)
	at org.junit.Assert.assertEquals(Assert.java:555)
	at org.junit.Assert.assertEquals(Assert.java:542)
	at plugins.WebOfTrust.util.TickerDelayedBackgroundJobTest
	   .testTriggerDefault(TickerDelayedBackgroundJobTest.java:210)

This would happen below the modified thread start loop (for(Thread t :
threads) t.start();). The reason was that Thread.start() can take up to
10ms for me, so the whole loop would be 100ms. This would exceed the
delay for the trigger to execute, and thus break the assert because it
assumed that it had not executed yet.