TickerDelayedBackgroundJobTest: Fix bogus timing
The reason why this was wrong is the same as in the previous
commit, please read its message first.
This in theory should have been in the previous commit, but was done
in a separate one because it is the only sleep time which is changed as
part of this  changeset. The previous commit did not change timings. So
I am putting this one into a separate commit.

(This timing was wrong because I had added it, using my false
assumptions. The others were right because they came from bertm)