TickerDelayed...Test.testWait...(): Fix possible wrong catch
We're interested in the InterruptedException by the second sleep, NOT in
the one from the first.
The first one should never throw InterruptedException as terminate() is
only called *after* it.
By using Sleeper, we ensure that InterruptedException is not thrown - it
converts it into RuntimeException.
Also, by reducing the size of the try-catch block to the part which it
wants to catch, we make it very clear what we're interested in.