TickerDelayedBackgroundJobTest: Remove warmup(), to be replaced
As far as I understand it, this function aimed to trigger JIT
compilation for causing time critical functions to be fast.
It failed to do so though in my experience, the tests would often fail
for me due to the timing requirements not being met.
This is likely because the function does not cover all code paths of
this test class, and also because it doesn't do enough iterations to
even trigger JIT.

I will replace this with a proper implementation for all code paths in
the following commits.