SynchronousDelayedBackgroundJobTest: Scale thread count with processor count
Instead of using arbitrary constant.
Use at least 2 threads instead of 1 because the tests aim to provoke
concurrency issues.
Notice: The substraction of 1 from the processor count is to leave a
CPU available for the thread which is creating the other threads.
Notice: perThreadExecutionCount is increased to an amount which causes a
maximum of ~ 1 second execution time per thread. This can be done since
we don't create more threads than processors of today's average desktop
(2-4 processors) anymore.