Add package plugins.WebOfTrust.util.jobs
The BackgroundJob classes will be moved to there in following commits.
This commit introduces the package by creating a package-info.java
there. The JavaDoc is intentionally written in a non-WebOfTrust-specific
way as I think it is likely that those classes will be moved to a
library which is used by more than just WOT.