SubscriptionManager: Replace TrivialTicker with DelayedBackgroundJob
Using TrivialTicker here was wrong: It could cause run() to execute
twice in parallel. The class was designed assuming this would NOT
happen, it is neither needed nor tolerable. DelayedBackgroundJob was
implemented specifically to replace such misuse of TrivialTicker, so it
is the right thing to use here.

Also, the complex shutdown logic is replaced by DelayedBackgroundJob
having terminate() / waitForTermination(), which behave as it was
requested in the comment in the complex shutdown logic.