IdentityFetcher: Document existing synchronization requirements
Notice: The other version of this function requires synchronization on
the WebOfTrust as can be seen in the previous commit. This is NOT needed
here: The function does not consume an Identity object but merely a
String with the ID of the identity and does not query stuff from the
WOT, so it is not in the synchronizationdomain of the WebOfTrust anymore.

This is how I had intended synchronization to work when I wrote the
code.
I have checked all callers, they either follow the requirements, or
by their JavaDoc require their callers to do.

(Doing this now since I am about to refactor startup/shutdown
synchronzation to integrate bertm's background jobs, and I need to be
sure about what the current state of multithreading is. This
investigation is especially indicated by the following TODO at the
class-level JavaDoc:
" * TODO: There is some synchronization missing for the fetcher in some
places where fetcher commands are issued.
 * It still works because those places are typically synchronized on the
WoT anyway. We should fix it nevertheless.")