IdentityFetcher: Replace TrivialTicker with DelayedBackgroundJob
Using TrivialTicker here was wrong: It could cause run() to execute
twice in parallel. The class was designed assuming this would NOT
happen, it is neither needed nor tolerable (for performance reasons).
DelayedBackgroundJob was implemented specifically to replace such
misuse of TrivialTicker, so it is the right thing to use here.

Notice: This is based upon similar code which was recently added to
SubscriptionManager