Copy class TransferThread from fred to WOT
This class was copy-pasted from fred class
freenet.support.TransferThread of commit
93c9564237a890f4871f0a3909f3276a72b3d769.

See
https://github.com/freenet/fred/blob/93c9564237a890f4871f0a3909f3276a72b3d769/src/freenet/support/TransferThread.java

This is necessary because class TransferThread contains bugs which are
to be fixed by this branch using the classes plugins.WebOfTrust.jobs.*
which are not present in fred. See
https://bugs.freenetproject.org/view.php?id=6423
https://bugs.freenetproject.org/view.php?id=6455
for the reasons.

I could have filed a pull request to put the classes needed for the fix
into fred, but class TransferThread is not used by fred and is an ugly
hack anyway. It should probably be removed from fred.