TransferThread: Replace TrivialTicker with DelayedBackgroundJob
Using TrivialTicker here was wrong: It could cause run() to execute
twice in parallel. The class was designed assuming this would NOT
happen, it is neither needed nor tolerable for thread safety.
DelayedBackgroundJob was implemented specifically to replace such
misuse of TrivialTicker, so it is the right thing to use here.

Notice: This is based upon similar code which was recently added to
IdentityFetcher, SubscriptionManager and
FCPClientReferenceImplementation.

Additional included fixes:
- This also fixes IntroductionClient from
possibly creating transfers at fred during termination of its parent
TransferThreads, which could cause them to be never cancelled. The
reason this could happen was that IntroductionClient created transfers
in event handlers which are called from fred, and TransferThread cannot
prevent those callbacks to happen during its shutdown. It is fixed by
changing IntroductionClient to only start transfers during iterate().
- IntroductionClient.interate() is fixed to not swallow scheduled
iterations if the minimal delay was not met. It now re-schedules itself
to after the minimal delay.
- IntroductionClient.solvePuzzle() and
XMLTransformer.importIntroductionPuzzle() are fixed to not
keep references to IntroductionPuzzle objects stored inside db4o outside
of a synchronized(IntroductionPuzzleStore) block. This is needed as a
workaround until https://bugs.freenetproject.org/view.php?id=6247 is
fixed.