IdentityInserter: Increase insert delay constants
- Increase thread period from 10 minutes to 1.5 hours. This fixes
inserts never completing because the IdentityInserter aborts
all inserts at each iteration (I know this is ugly and should be fixed,
but we want to get the release done ASAP). Also, define the thread
period via  the maximal insert delay since it doesn't make sense to have
a higher period than the maximum time we want inserts to be delayed for.
- Increase maximal insert delay as well, because the thread period is
computed from this, and  because this will reduce general load on the
network.
- Increase minimal insert delay to reduce load on the network.

To compensate for the higher constants, cause a thread iteration at
WebOfTrust.createOwnIdentity(). This makes sense anyway, since the
computations at the thread iteration of IdentityInserter will
always allow an immediate insert for newly created identities,
independent from minimal insert delay.

The increased thread period fixes issue
https://bugs.freenetproject.org/view.php?id=6447