Change comment to code quality TODO
There is no bugtracker entry for this, so its better to add a TODO at
least.
Also improve its wording a bit.

(The previous commit said I would keep the bugtracker entry open, but I
did not do so for the sake of having it in the changelog in the next
release)