IntroductionClient: Add workarounds for issue 0006247
https://bugs.freenetproject.org/view.php?id=6247

By not keeping hot references to objects of the WebOfTrust outside a
synchronized(WebOfTrust) block, we ensure that they can always be GCed
by the rollback() code to enforce that they are refreshed in
transactions which happen after the rollback.
Yes, this is ugly. But the real fix to 6247 would take months.