SubscriptionManager: Also start() in unit tests
(Notice: Unlike expected, this does not fix the current unit test
failures. I am committing it nervetheless because it is a good idea to
have this for possible future unit tests)

We didn't start it previously since that would cause its thread to
iterate. But this isn't the case anymore, start() has recently been
changed to not do anything during unit tests.

Also, SubscriptionManager.start() is changed by this commit to not do
anything in a similar way as IdentityFetcher.start() does:
Create a fake Runnable which won't do any actual work; and shove it into
a Ticker/Executor of our own since we cannot obtain them from the
PluginRespirator as there is none in unit tests.
This is better than the previous solution of leaving mJob be a
MockDelayedBackgroundJob because that was the same state as if start()
had never been called. By now changing mJob to a real
DelayedBackgroundJob, we mark the SubscriptionManager as started, which
can be valuable information to certain tests. For example it allows
triggerExecution() to detect that it was called before start().