SubscriptionManager: Fix some unit tests failing due to "Should not be called  before start() or after stop()"
SubscriptionManagerFCPTest.testSubscribeUnsubscribe() would
fail with:
[junit] Testcase: testSubscribeUnsubscribe took 14.278 sec
[junit]     FAILED
[junit] expected:<true> but was:<false>
[junit] junit.framework.AssertionFailedError: expected:<true> but
was:<false>
[junit]     at
plugins.WebOfTrust.SubscriptionManagerFCPTest.testSubscribeTo(SubscriptionManagerFCPTest.java:203)
[junit]     at
plugins.WebOfTrust.SubscriptionManagerFCPTest.testSubscribeUnsubscribe(SubscriptionManagerFCPTest.java:159)
[junit]     at
plugins.WebOfTrust.SubscriptionManagerFCPTest.testSubscribeUnsubscribe(SubscriptionManagerFCPTest.java:150)

The fix is to make scheduleNotificationProcessing() NOT assert() that
the SubscriptionManager was not stop()ed yet. The unit tests
intentionally stop() it, see AbstractFullNodeTest.setUpNode().
To make this fix possible, the class is modified to have its
start()/stop() and scheduleNotificationProcessing() to work the same way
as the similar functions in IdentityFetcher
(scheduleNotificationProcessing is called scheduleCommandProcessing
there). Particularly, the possible states of the mJob variable are
changed to clearly distinguish between the three states "start() not
called  yet", "start() called", "stop() called". This allows us to keep
an "assert(!(start not called))" in scheduleNotificationProcessing()
instead of only having an "assert(!(start() not called OR stop not
called))". So we can preserve some assert-safety.

This fix is only for class SubscriptionManager. Similar fixes are to
follow for TransferThread.