(Own)Identity / Trust / Score: Make toString() output everything which equals() compares
This will be very useful when debugging the output of DebugFCPClient:
Its purpose is to catch mismatches of (Own)Identity/Trust/Score objects.
It already uses the toString() functions to log the mismatching objects,
but the toString() didn't contain the values of all member variables so
it was impossible to tell WHAT is mismatching.

Notice:
- The toString() of Trust/Score did already contain most of the member
variables, if not all. I've reformatted the output nevertheless so it
matches what (Own)Identity does.
- Synchronization of toString() is removed in both Trust and Score. We
always synchronize upon the WebOfTrust when consistency of such objects
is needed, there is no need for internal synchronization in them.
- The truster and trustee are NOT added to the String of Trust/Score
anymore because their ID is already fully contained in mID. When
analyzing logs, one will typically search them for identity IDs and not
for  things such as nicknames. Also, equals() does not care about
truster/trustee, it only cares about mID as well.
- Not fixing this for (Own)IntroductionPuzzle yet since they are not
wired into event-notifications yet, and thus are not affected by what
DebugFCPClient does. Adding a TODO to them nevertheless so this doesn't
get forgotten about.