XMLTransformer: Fix import of unwanted identities
Notice that this was not a severe security problem:
As part of its design goals, the Score computation algorithm will ignore
Trusts of Identitys for which shouldFetchIdentity() is false. So the
unwanted Identitys did only bloat the database a bit, their Trusts were
not able to affect the Score view of any OwnIdentity.