(Own)IdentityTest.testClone(): Fix to show failure of preserving mLastChangedDate
This fixes the said unit tests to demonstrate that (Own)Identity.clone()
does not preserve mLastChangedDate. The tests will now fail. This is
intentional and will be fixed in the following commit.

The failure of preserving the date only happens if an (Own)Identity has
a context/property, and the tests failed to add a context/property. This
also implicitly improves the tests to check whether clone() preserves
contexts/properties.

The new failures of the tests will be similar to:

junit.framework.AssertionFailedError: protected java.util.Date
plugins.WebOfTrust.Identity.mLastChangedDate expected:<Tue Mar 10
05:48:36 CET 2015> but was:<Tue Mar 10 05:48:36 CET 2015>
	at junit.framework.Assert.fail(Assert.java:47)
	at junit.framework.Assert.failNotEquals(Assert.java:280)
	at junit.framework.Assert.assertEquals(Assert.java:64)
	at plugins.WebOfTrust.AbstractJUnit3BaseTest
        .testClone(AbstractJUnit3BaseTest.java:117)
	at plugins.WebOfTrust.OwnIdentityTest
        .testClone(OwnIdentityTest.java:98)